pyproject.toml
[project]
name = "ai-langgraph-react-agentexecutor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "boto3>=1.40.0",
    "fastmcp>=2.10.6",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-aws>=0.2.30",
    "langchain-core>=0.3.72",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-ollama>=0.3.6",
    "langgraph>=0.6.2",
    "mcp[cli]>=1.12.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "react>=4.3.0",
    "serpapi>=0.1.5",
    "streamlit>=1.47.1",
]
